+++ src/output/saveFileManager/SaveFileManager.cpp
@@ -104,11 +104,13 @@ void SaveFileManager::OnSaveFileCreated()
     ClearEventFlag(static_cast<EventFlags>(FLAG_MILES_SS_STATION_BACK));
     ClearEventFlag(static_cast<EventFlags>(FLAG_KNUCKLES_SS_STATION_BACK));
     ClearEventFlag(static_cast<EventFlags>(FLAG_AMY_SS_STATION_BACK));
+    
     //Casino-Hotel door
     ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_SS_HOTEL_BACK));
     ClearEventFlag(static_cast<EventFlags>(FLAG_MILES_SS_HOTEL_BACK));
     ClearEventFlag(static_cast<EventFlags>(FLAG_KNUCKLES_SS_HOTEL_BACK));
     ClearEventFlag(static_cast<EventFlags>(FLAG_AMY_SS_HOTEL_BACK));
+    
     //Casino Door
     ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_SS_ENTRANCE_CASINO));
     ClearEventFlag(static_cast<EventFlags>(FLAG_MILES_SS_ENTRANCE_CASINO));